pyproject.toml
[tool.poetry]
name = "rinhack-kanban-service"
version = "0.1.0"
description = ""
authors = ["JKearnsl <boyarshinovn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
SQLAlchemy = "^2.0.19"
asyncpg = "^0.29.0"

aiohttp = "3.9.0"
aiobotocore = "^2.7.0"

python-consul = "^1.1.0"
pyjwt = "^2.8.0"
grpcio = "^1.59.0"
APScheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.7"
alembic = "^1.12.0"
grpcio-tools = "^1.57.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"